+++ HISTORY.md
@@ -1,5 +1,11 @@
 T2WML API History
 ===================================
+Changes in version 0.0.12:
+------------------------
+* rename add_nodes_from_file to add_entities_from_file
+* rename .t2wmlproj to project.t2wml, so that it can be viewed on max and linux
+
+
 Changes in version 0.0.11:
 ------------------------
 * rename add_wikidata_file to more accurate add_entity_file